Create a_star.py
I added an example of python work I have done.  In this particular example the parts I wrote were the ones that dealt with the a star algorithm itself.  I did not write the part that reads in the input file to create the obstacles or the part that checks line intersections.